+++ openghg_inversions/models/__init__.py
@@ -8,11 +8,13 @@ configure_pytensor()
 
 from openghg_inversions.models.components import (
     LinearComponentResult,
+    StateLinearComponentResult,
     add_inferpymc_likelihood_component,
     add_linear_component,
     add_model_data,
     add_offset_component,
     add_sigma_component,
+    add_state_linear_component,
 )
 from openghg_inversions.models.coords import (
     CoordRegistry,
@@ -22,6 +24,12 @@ from openghg_inversions.models.coords import (
     restore_inferencedata_coords,
 )
 from openghg_inversions.models.priors import parse_prior
+from openghg_inversions.models.state_activity import (
+    ResolvedStateActivity,
+    StateActivity,
+    active_prior_args,
+    resolve_state_activity,
+)
 from openghg_inversions.models.rhime import (
     DEFAULT_BC_PRIOR,
     DEFAULT_OFFSET_PRIOR,
@@ -43,15 +51,19 @@ __all__ = [
     "DEFAULT_SIGMA_PRIOR",
     "DEFAULT_X_PRIOR",
     "RhimeModelSpec",
+    "ResolvedStateActivity",
     "SectorSpec",
+    "StateActivity",
     "add_coords",
     "attach_coord_registry",
     "get_coord_registry",
     "restore_inferencedata_coords",
     "parse_prior",
     "LinearComponentResult",
+    "StateLinearComponentResult",
     "add_model_data",
     "add_linear_component",
+    "add_state_linear_component",
     "add_sigma_component",
     "add_offset_component",
     "add_inferpymc_likelihood_component",
@@ -60,4 +72,6 @@ __all__ = [
     "build_rhime_multisector_model",
     "build_rhime_multisector_model_from_spec",
     "safe_pymc_name",
+    "active_prior_args",
+    "resolve_state_activity",
 ]